# swagger_server/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: proVis
  description: API documentation for proVis website
  contact:
    email: aavegj1904@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
- url: https://proVis.herokuapp.com/api/
...
paths:
...
  /user/emailRegister:
    post:
      tags:
      - User
      summary: registration
      description: ""
      operationId: email_register
      requestBody:
        description: registers the user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        "400":
          description: Invalid username/email/password supplied
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/emailLogin:
    post:
      tags:
      - User
      summary: email login
      description: ""
      operationId: email_login
      requestBody:
        description: logs in the user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        "400":
          description: Invalid username/email/password supplied
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/sendVerificationMail:
    get:
      tags:
      - User
      summary: verification email
      description: ""
      operationId: send_verification_mail
      parameters:
      - name: Authorization
        in: header
        description: an authorization header
        required: true
        style: simple
        explode: false
        schema:
          type: string
          example: Bearer 85jkl
      responses:
        "200":
          description: successful operation
      x-openapi-router-controller: swagger_server.controllers.user_controller
  /user/verifyEmail:
    post:
      tags:
      - User
      summary: verify email
      description: ""
      operationId: verify_email
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: success
                x-content-type: application/json
        "400":
          description: Invalid username/email/password supplied
      x-openapi-router-controller: swagger_server.controllers.user_controller
components:
  schemas:
...
    UserRegister:
      type: object
      properties:
        name:
          type: string
          example: aaveg
        email_id:
          type: string
          example: aavegj1904gmail.com
        password:
          type: string
          example: "12345"
    UserLogin:
      type: object
      properties:
        email_id:
          type: string
          example: aavegj1904gmail.com
        password:
          type: string
          example: "12345"
    RegisterResponse:
      type: object
      properties:
        token:
          type: string
          example: 38484jfnl
        info:
          $ref: '#/components/schemas/RegisterResponse_info'
      example:
        token: 38484jfnl
        info:
          cus_uid: 343c
          username: aaveg
    VerifyEmail:
      type: object
      properties:
        token:
          type: string
          example: 38484jfnl
        user_id:
          type: string
...
    RegisterResponse_info:
      type: object
      properties:
        cus_uid:
          type: string
          example: 343c
        username:
          type: string
          example: aaveg
      example:
        cus_uid: 343c
        username: aaveg